Change internal terminology and make use of go-github internals
Having two concepts of "PullRequest" was getting confusing. While not
perfect "ReviewRequest" conveys the idea that we're managing the process
of someone reviewing a pull request.

Also got rid of some crazy stupid regex stuff to make use of the
underlying github api client tools that go-github exposes. It's not
pretty at the moment, but it's better than parsing URLs with regex!